+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- salesforce picklist dynamic options
+
 ### Fixed
 
 - github action permissions to allow docs and github release jobs